restructure table folders and fix lens table styling
