+++ src/js/engagement_view/src/components/engagementView/sidebar/tables/lensTable/selectLens.tsx
@@ -4,7 +4,7 @@ import Button from "@material-ui/core/Button";
 import TableCell from "@material-ui/core/TableCell";
 import TableRow from "@material-ui/core/TableRow";
 
-import { useStyles } from './styles';
+import { useStyles } from './lensTableStyles';
 
 import { SelectLensProps } from "types/LensAndNodeTableTypes"
