+++ src/js/engagement_view/src/components/engagementView/sidebar/tables/nodeTable/nodeTableStyles.tsx
