+++ src/js/engagement_view/src/components/engagementView/sidebar/tables/toggleNodeDetailTable.tsx
@@ -7,7 +7,7 @@ import { ToggleNodeTableProps } from "types/LensAndNodeTableTypes";
 
 import { NodeDetails } from "../LensAndNodeTableContainer";
 
-import { useStyles } from "./lensTable/styles";
+import { useStyles } from "./lensTable/lensTableStyles";
 
 
 export function ToggleNodeDetailTable({ curNode }: ToggleNodeTableProps) {
@@ -19,7 +19,7 @@ export function ToggleNodeDetailTable({ curNode }: ToggleNodeTableProps) {
 				<div className={classes.header}>
 					<b className={classes.title}> Node Details</b>
 					<Button
-						className={classes.button}
+						className={classes.lensToggleBtn}
 						onClick={() => {
 							setToggle((toggled) => !toggled);
 						}}
